Chop, chop, chop.
